@@ -1,14 +1,14 @@
 [build-system]
-requires = ["pdm-pep517"]
-build-backend = "pdm.pep517.api"
+requires = ["pdm-backend"]
+build-backend = "pdm.backend"
 
 [project]
 name = "aria2p"
 description = "Command-line tool and library to interact with an aria2c daemon process with JSON-RPC."
 authors = [{name = "Timothée Mazzucotelli", email = "pawamoy@pm.me"}]
-license = "ISC"
+license = {text = "ISC"}
 readme = "README.md"
-requires-python = ">=3.7"
+requires-python = ">=3.8"
 keywords = ["aria2", "aria2c", "aria2-cli"]
 dynamic = ["version"]
 classifiers = [
@@ -17,12 +17,11 @@ classifiers = [
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.6",
-    "Programming Language :: Python :: 3.7",
     "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
     "Topic :: Utilities",
     "Typing :: Typed",
 ]
@@ -55,51 +54,39 @@ aria2p = "aria2p.cli.main:main"
 
 [tool.pdm]
 version = {source = "scm"}
+plugins = [
+    "pdm-multirun",
+]
 
 [tool.pdm.build]
 package-dir = "src"
 editable-backend = "editables"
 
 [tool.pdm.dev-dependencies]
-duty = ["duty>=0.7"]
+duty = ["duty>=0.10"]
+ci-quality = ["aria2p[duty,docs,quality,typing,security]"]
+ci-tests = ["aria2p[duty,tests]"]
 docs = [
-    "mkdocs>=1.3",
-    "mkdocs-coverage>=0.2",
-    "mkdocs-gen-files>=0.3",
-    "mkdocs-literate-nav>=0.4",
-    "mkdocs-material>=7.3",
-    "mkdocs-section-index>=0.3",
-    "mkdocstrings[python]>=0.18",
+    "black>=23.1",
     "markdown-callouts>=0.2",
     "markdown-exec>=0.5",
+    "mkdocs>=1.5",
+    "mkdocs-coverage>=0.2",
+    "mkdocs-gen-files>=0.3",
+    "mkdocs-git-committers-plugin-2>=1.1",
+    "mkdocs-literate-nav>=0.4",
+    "mkdocs-material>=7.3",
+    "mkdocs-minify-plugin>=0.6.4",
+    "mkdocstrings[python]>=0.18",
     "toml>=0.10",
 ]
-format = [
-    "autoflake>=1.4",
-    "black>=21.10b0",
-    "isort>=5.10",
-]
 maintain = [
-    "git-changelog>=0.4",
+    "black>=23.1",
+    "blacken-docs>=1.13",
+    "git-changelog>=1.0",
 ]
 quality = [
-    # TODO: remove once importlib-metadata version conflict is resolved
-    "importlib-metadata<5; python_version < '3.8'",
-    "flake8>=4; python_version >= '3.8'",
-
-    "darglint>=1.8",
-    "flake8-bandit>=2.1",
-    "flake8-black>=0.2",
-    "flake8-bugbear>=21.9",
-    "flake8-builtins>=1.5",
-    "flake8-comprehensions>=3.7",
-    "flake8-docstrings>=1.6",
-    "flake8-pytest-style>=1.5",
-    "flake8-string-format>=0.3",
-    "flake8-tidy-imports>=4.5",
-    "flake8-variables-names>=0.0",
-    "pep8-naming>=0.12",
-    "wps-light>=0.15",
+    "ruff>=0.0.246",
 ]
 tests = [
     "fastapi~=0.61",
@@ -114,23 +101,11 @@ tests = [
 typing = [
     "mypy>=0.910",
     "types-markdown>=3.3",
+    "types-pyyaml>=6.0",
     "types-toml>=0.10",
     "types-setuptools",
     "types-requests",
 ]
-security = ["safety>=2"]
-
-[tool.black]
-line-length = 120
-exclude = "tests/fixtures"
-
-[tool.isort]
-line_length = 120
-not_skip = "__init__.py"
-multi_line_output = 3
-force_single_line = false
-balanced_wrapping = true
-default_section = "THIRDPARTY"
-known_first_party = "aria2p"
-include_trailing_comma = true
-
+security = [
+    "safety>=2",
+]
