@@ -1,3 +1,5 @@
+"""Port-related utilities."""
+
 import re
 import sys
 from collections import defaultdict
@@ -6,7 +8,7 @@ from os.path import isfile, join
 from pathlib import Path
 
 
-def get_ports():
+def get_ports() -> dict[int, list[int]]:  # noqa: D103
     port_regex = re.compile(r"port=[0-9]{4}")
     test_files = [Path("tests") / f for f in listdir("tests") if isfile(join("tests", f))]
     used_ports = defaultdict(list)
@@ -20,7 +22,7 @@ def get_ports():
     return used_ports
 
 
-def check(ports):
+def check(ports: dict[int, list[int]]) -> int:  # noqa: D103
     check_ok = True
     for port, matches in ports.items():
         if port in blacklisted_ports:
@@ -39,9 +41,9 @@ def check(ports):
     return 0 if check_ok else 1
 
 
-def get_unused(ports):
+def get_unused(ports: dict[int, list[int]]) -> list[int]:  # noqa: D103
     finalists = []
-    sorted_ports = list(sorted(ports.keys()))
+    sorted_ports = sorted(ports.keys())
     following = False
     for candidate in range(sorted_ports[0], sorted_ports[-1] + 2):
         if candidate not in sorted_ports:
@@ -53,7 +55,7 @@ def get_unused(ports):
     return finalists
 
 
-def next_unused(port):
+def next_unused(port: int) -> int | None:  # noqa: D103
     for next_port in unused_ports:
         if next_port > port:
             return next_port
