@@ -1,5 +1,4 @@
-"""
-aria2p package.
+"""aria2p package.
 
 Command-line tool and library to interact with an aria2c daemon process with JSON-RPC.
 """
@@ -7,6 +6,7 @@ Command-line tool and library to interact with an aria2c daemon process with JSO
 from __future__ import annotations
 
 import sys
+from typing import TextIO
 
 from loguru import logger
 
@@ -19,13 +19,12 @@ from aria2p.stats import Stats
 logger.disable("aria2p")
 
 
-def enable_logger(sink=sys.stderr, level="WARNING"):
-    """
-    Enable the logging of messages.
+def enable_logger(sink: str | TextIO = sys.stderr, level: str = "WARNING") -> None:
+    """Enable the logging of messages.
 
     Configure the `logger` variable imported from `loguru`.
 
-    Arguments:
+    Parameters:
         sink (file): An opened file pointer, or stream handler. Default to standard error.
         level (str): The log level to use. Possible values are TRACE, DEBUG, INFO, WARNING, ERROR, CRITICAL.
             Default to WARNING.
