@@ -1,5 +1,4 @@
-"""
-Aria2 API.
+"""Aria2 API.
 
 This module defines the API class, which makes use of a JSON-RPC client to provide higher-level methods to
 interact easily with a remote aria2c process.
@@ -12,7 +11,7 @@ import shutil
 import threading
 from base64 import b64encode
 from pathlib import Path
-from typing import Callable, Dict, Iterator, List, TextIO, Tuple, Union
+from typing import TYPE_CHECKING, Callable, Dict, Iterator, List, TextIO, Tuple, Union
 
 from loguru import logger
 
@@ -20,7 +19,9 @@ from aria2p.client import Client, ClientException
 from aria2p.downloads import Download
 from aria2p.options import Options
 from aria2p.stats import Stats
-from aria2p.types import PathOrStr
+
+if TYPE_CHECKING:
+    from aria2p.types import PathOrStr
 
 OptionsType = Union[Options, dict]
 OperationResult = Union[bool, ClientException]
@@ -28,8 +29,7 @@ InputFileContentsType = List[Tuple[List[str], Dict[str, str]]]
 
 
 class API:
-    """
-    A class providing high-level methods to interact with a remote aria2c process.
+    """A class providing high-level methods to interact with a remote aria2c process.
 
     This class is instantiated with a reference to a [`Client`][aria2p.client.Client] instance. It then uses this client
     to call remote procedures, or remote methods. While the client methods reflect exactly what aria2c is providing
@@ -39,10 +39,9 @@ class API:
     """
 
     def __init__(self, client: Client | None = None) -> None:
-        """
-        Initialize the object.
+        """Initialize the object.
 
-        Arguments:
+        Parameters:
             client: An instance of the [aria2p.client.Client][] class.
         """
         self.client = client or Client()
@@ -51,19 +50,18 @@ class API:
     def __repr__(self) -> str:
         return f"API({self.client!r})"
 
-    def add(  # noqa: WPS231 (not that complex)
+    def add(
         self,
         uri: str,
         options: OptionsType | None = None,
         position: int | None = None,
     ) -> list[Download]:
-        """
-        Add a download (guess its type).
+        """Add a download (guess its type).
 
         If the provided URI is in fact a file-path, and is neither a torrent or a metalink,
         then we read its lines and try to add each line as a download, recursively.
 
-        Arguments:
+        Parameters:
             uri: The URI or file-path to add.
             options: An instance of the [`Options`][aria2p.options.Options] class or a dictionary
                 containing aria2c options to create the download with.
@@ -102,10 +100,9 @@ class API:
         return new_downloads
 
     def add_magnet(self, magnet_uri: str, options: OptionsType | None = None, position: int | None = None) -> Download:
-        """
-        Add a download with a Magnet URI.
+        """Add a download with a Magnet URI.
 
-        Arguments:
+        Parameters:
             magnet_uri: The Magnet URI.
             options: An instance of the [`Options`][aria2p.options.Options] class or a dictionary
                 containing aria2c options to create the download with.
@@ -117,10 +114,7 @@ class API:
         if options is None:
             options = {}
 
-        if isinstance(options, Options):
-            client_options = options.get_struct()
-        else:
-            client_options = options
+        client_options = options.get_struct() if isinstance(options, Options) else options
 
         gid = self.client.add_uri([magnet_uri], client_options, position)
 
@@ -133,10 +127,9 @@ class API:
         options: OptionsType | None = None,
         position: int | None = None,
     ) -> Download:
-        """
-        Add a download with a torrent file (usually .torrent extension).
+        """Add a download with a torrent file (usually .torrent extension).
 
-        Arguments:
+        Parameters:
             torrent_file_path: The path to the torrent file.
             uris: A list of URIs used for Web-seeding.
             options: An instance of the [`Options`][aria2p.options.Options] class or a dictionary
@@ -152,10 +145,7 @@ class API:
         if options is None:
             options = {}
 
-        if isinstance(options, Options):
-            client_options = options.get_struct()
-        else:
-            client_options = options
+        client_options = options.get_struct() if isinstance(options, Options) else options
 
         with open(torrent_file_path, "rb") as stream:
             torrent_contents = stream.read()
@@ -171,10 +161,9 @@ class API:
         options: OptionsType | None = None,
         position: int | None = None,
     ) -> list[Download]:
-        """
-        Add a download with a Metalink file.
+        """Add a download with a Metalink file.
 
-        Arguments:
+        Parameters:
             metalink_file_path: The path to the Metalink file.
             options: An instance of the [`Options`][aria2p.options.Options] class or a dictionary
                 containing aria2c options to create the download with.
@@ -186,10 +175,7 @@ class API:
         if options is None:
             options = {}
 
-        if isinstance(options, Options):
-            client_options = options.get_struct()
-        else:
-            client_options = options
+        client_options = options.get_struct() if isinstance(options, Options) else options
 
         with open(metalink_file_path, "rb") as stream:
             metalink_contents = stream.read()
@@ -205,10 +191,9 @@ class API:
         options: OptionsType | None = None,
         position: int | None = None,
     ) -> Download:
-        """
-        Add a download with a URL (or more).
+        """Add a download with a URL (or more).
 
-        Arguments:
+        Parameters:
             uris: A list of URIs that point to the same resource.
             options: An instance of the `Options` class or a dictionary
                 containing aria2c options to create the download with.
@@ -221,22 +206,18 @@ class API:
         if options is None:
             options = {}
 
-        if isinstance(options, Options):
-            client_options = options.get_struct()
-        else:
-            client_options = options
+        client_options = options.get_struct() if isinstance(options, Options) else options
 
         gid = self.client.add_uri(uris, client_options, position)
 
         return self.get_download(gid)
 
     def search(self, patterns: list[str]) -> list[Download]:
-        """
-        Not implemented.
+        """Not implemented.
 
         Search and return [`Download`][aria2p.downloads.Download] objects based on multiple patterns.
 
-        Arguments:
+        Parameters:
             patterns: The patterns used to filter the download list.
 
         Raises:
@@ -275,10 +256,9 @@ class API:
         raise NotImplementedError
 
     def get_download(self, gid: str) -> Download:
-        """
-        Get a [`Download`][aria2p.downloads.Download] object thanks to its GID.
+        """Get a [`Download`][aria2p.downloads.Download] object thanks to its GID.
 
-        Arguments:
+        Parameters:
             gid: The GID of the download to get.
 
         Returns:
@@ -287,10 +267,9 @@ class API:
         return Download(self, self.client.tell_status(gid))
 
     def get_downloads(self, gids: list[str] | None = None) -> list[Download]:
-        """
-        Get a list of [`Download`][aria2p.downloads.Download] object thanks to their GIDs.
+        """Get a list of [`Download`][aria2p.downloads.Download] object thanks to their GIDs.
 
-        Arguments:
+        Parameters:
             gids: The GIDs of the downloads to get. If None, return all the downloads.
 
         Returns:
@@ -311,10 +290,9 @@ class API:
         return downloads
 
     def move(self, download: Download, pos: int) -> int:
-        """
-        Move a download in the queue, relatively to its current position.
+        """Move a download in the queue, relatively to its current position.
 
-        Arguments:
+        Parameters:
             download: The download object to move.
             pos: The relative position (1 to move down, -1 to move up, -2 to move up two times, etc.).
 
@@ -324,10 +302,9 @@ class API:
         return self.client.change_position(download.gid, pos, "POS_CUR")
 
     def move_to(self, download: Download, pos: int) -> int:
-        """
-        Move a download in the queue, with absolute positioning.
+        """Move a download in the queue, with absolute positioning.
 
-        Arguments:
+        Parameters:
             download: The download object to move.
             pos: The absolute position in the queue where to move the download. 0 for top, -1 for bottom.
 
@@ -336,16 +313,15 @@ class API:
         """
         if pos < 0:
             how = "POS_END"
-            pos = -pos  # noqa: WPS434
+            pos = -pos
         else:
             how = "POS_SET"
         return self.client.change_position(download.gid, pos, how)
 
     def move_up(self, download: Download, pos: int = 1) -> int:
-        """
-        Move a download up in the queue.
+        """Move a download up in the queue.
 
-        Arguments:
+        Parameters:
             download: The download object to move.
             pos: Number of times to move up. With negative values, will move down (use move or move_down instead).
 
@@ -355,10 +331,9 @@ class API:
         return self.client.change_position(download.gid, -pos, "POS_CUR")
 
     def move_down(self, download: Download, pos: int = 1) -> int:
-        """
-        Move a download down in the queue.
+        """Move a download down in the queue.
 
-        Arguments:
+        Parameters:
             download: The download object to move.
             pos: Number of times to move down. With negative values, will move up (use move or move_up instead).
 
@@ -368,10 +343,9 @@ class API:
         return self.client.change_position(download.gid, pos, "POS_CUR")
 
     def move_to_top(self, download: Download) -> int:
-        """
-        Move a download to the top of the queue.
+        """Move a download to the top of the queue.
 
-        Arguments:
+        Parameters:
             download: The download object to move.
 
         Returns:
@@ -380,10 +354,9 @@ class API:
         return self.client.change_position(download.gid, 0, "POS_SET")
 
     def move_to_bottom(self, download: Download) -> int:
-        """
-        Move a download to the bottom of the queue.
+        """Move a download to the bottom of the queue.
 
-        Arguments:
+        Parameters:
             download: The download object to move.
 
         Returns:
@@ -391,15 +364,14 @@ class API:
         """
         return self.client.change_position(download.gid, 0, "POS_END")
 
-    def retry_downloads(  # noqa: WPS231 (not that complex)
+    def retry_downloads(
         self,
         downloads: list[Download],
-        clean: bool = False,
+        clean: bool = False,  # noqa: FBT001,FBT002
     ) -> list[OperationResult]:
-        """
-        Resume failed downloads from where they left off with new GIDs.
+        """Resume failed downloads from where they left off with new GIDs.
 
-        Arguments:
+        Parameters:
             downloads: The list of downloads to remove.
             clean: Whether to remove the aria2 control file as well.
 
@@ -412,7 +384,7 @@ class API:
             if not download.has_failed:
                 continue
             try:
-                uri = download.files[0].uris[0]["uri"]  # noqa: WPS219 (deep access)
+                uri = download.files[0].uris[0]["uri"]
             except IndexError:
                 continue
             try:
@@ -428,17 +400,16 @@ class API:
 
         return result
 
-    def remove(  # noqa: WPS231 (complex, maybe we could split it)
+    def remove(
         self,
         downloads: list[Download],
-        force: bool = False,
-        files: bool = False,
-        clean: bool = True,
+        force: bool = False,  # noqa: FBT001,FBT002
+        files: bool = False,  # noqa: FBT001,FBT002
+        clean: bool = True,  # noqa: FBT001,FBT002
     ) -> list[OperationResult]:
-        """
-        Remove the given downloads from the list.
+        """Remove the given downloads from the list.
 
-        Arguments:
+        Parameters:
             downloads: The list of downloads to remove.
             force: Whether to force the removal or not.
             files: Whether to remove downloads files as well.
@@ -448,10 +419,7 @@ class API:
             Success or failure of the operation for each given download.
         """
         # Note: batch/multicall candidate
-        if force:
-            remove_func = self.client.force_remove
-        else:
-            remove_func = self.client.remove
+        remove_func = self.client.force_remove if force else self.client.remove
 
         result: list[OperationResult] = []
 
@@ -470,46 +438,40 @@ class API:
                 logger.debug(f"Try to remove download {download.gid}")
                 try:
                     removed_gid = remove_func(download.gid)
-                except ClientException as error:  # noqa: WPS440 (block variable overlap)
+                except ClientException as error:
                     logger.exception(error)
                     result.append(error)
                 else:
                     logger.success(f"Removed download {download.gid}")
                     result.append(True)
-                    try:  # noqa: WPS505 (nested try)
+                    try:  # (nested try)
                         self.client.remove_download_result(download.gid)
-                    except ClientException as error:  # noqa: WPS440 (block variable overlap)
+                    except ClientException as error:
                         logger.debug(f"Failed to remove download result {download.gid}")
                         logger.opt(exception=True).trace(error)
                     if removed_gid != download.gid:
                         logger.debug(
                             f"Removed download GID#{removed_gid} is different than download GID#{download.gid}",
                         )
-                        try:  # noqa: WPS505 (nested try)
+                        try:
                             self.client.remove_download_result(removed_gid)
-                        except ClientException as error:  # noqa: WPS440 (block variable overlap)
+                        except ClientException as error:
                             logger.debug(f"Failed to remove download result {removed_gid}")
                             logger.opt(exception=True).trace(error)
 
             if clean:
-                # FUTURE: use missing_ok parameter on Python 3.8
-                try:
-                    download.control_file_path.unlink()
-                except FileNotFoundError:
-                    logger.debug(f"aria2 control file {download.control_file_path} was not found")
-                else:
-                    logger.debug(f"Removed control file {download.control_file_path}")
+                download.control_file_path.unlink(missing_ok=True)
+                logger.debug(f"Removed control file {download.control_file_path}")
 
             if files and result[-1]:
                 self.remove_files([download], force=True)
 
         return result
 
-    def remove_all(self, force: bool = False) -> bool:
-        """
-        Remove all downloads from the list.
+    def remove_all(self, force: bool = False) -> bool:  # noqa: FBT001,FBT002
+        """Remove all downloads from the list.
 
-        Arguments:
+        Parameters:
             force: Whether to force the removal or not.
 
         Returns:
@@ -517,11 +479,10 @@ class API:
         """
         return all(self.remove(self.get_downloads(), force=force))
 
-    def pause(self, downloads: list[Download], force: bool = False) -> list[OperationResult]:
-        """
-        Pause the given (active) downloads.
+    def pause(self, downloads: list[Download], force: bool = False) -> list[OperationResult]:  # noqa: FBT001,FBT002
+        """Pause the given (active) downloads.
 
-        Arguments:
+        Parameters:
             downloads: The list of downloads to pause.
             force: Whether to pause immediately without contacting servers or not.
 
@@ -529,10 +490,7 @@ class API:
             Success or failure of the operation for each given download.
         """
         # Note: batch/multicall candidate
-        if force:
-            pause_func = self.client.force_pause
-        else:
-            pause_func = self.client.pause
+        pause_func = self.client.force_pause if force else self.client.pause
 
         result: list[OperationResult] = []
 
@@ -548,27 +506,22 @@ class API:
 
         return result
 
-    def pause_all(self, force: bool = False) -> bool:
-        """
-        Pause all (active) downloads.
+    def pause_all(self, force: bool = False) -> bool:  # noqa: FBT001,FBT002
+        """Pause all (active) downloads.
 
-        Arguments:
+        Parameters:
             force: Whether to pause immediately without contacting servers or not.
 
         Returns:
             Success or failure of the operation to pause all downloads.
         """
-        if force:
-            pause_func = self.client.force_pause_all
-        else:
-            pause_func = self.client.pause_all
+        pause_func = self.client.force_pause_all if force else self.client.pause_all
         return pause_func() == "OK"
 
     def resume(self, downloads: list[Download]) -> list[OperationResult]:
-        """
-        Resume (unpause) the given downloads.
+        """Resume (unpause) the given downloads.
 
-        Arguments:
+        Parameters:
             downloads: The list of downloads to resume.
 
         Returns:
@@ -590,8 +543,7 @@ class API:
         return result
 
     def resume_all(self) -> bool:
-        """
-        Resume (unpause) all downloads.
+        """Resume (unpause) all downloads.
 
         Returns:
             Success or failure of the operation to resume all downloads.
@@ -599,8 +551,7 @@ class API:
         return self.client.unpause_all() == "OK"
 
     def purge(self) -> bool:
-        """
-        Purge completed, removed or failed downloads from the queue.
+        """Purge completed, removed or failed downloads from the queue.
 
         Returns:
             Success or failure of the operation.
@@ -608,8 +559,7 @@ class API:
         return self.client.purge_download_result() == "OK"
 
     def autopurge(self) -> bool:
-        """
-        Purge completed, removed or failed downloads from the queue.
+        """Purge completed, removed or failed downloads from the queue.
 
         Deprecated. Use [`purge`][aria2p.api.API.purge] instead.
 
@@ -620,10 +570,9 @@ class API:
         return self.purge()
 
     def get_options(self, downloads: list[Download]) -> list[Options]:
-        """
-        Get options for each of the given downloads.
+        """Get options for each of the given downloads.
 
-        Arguments:
+        Parameters:
             downloads: The list of downloads to get the options of.
 
         Returns:
@@ -636,8 +585,7 @@ class API:
         return options
 
     def get_global_options(self) -> Options:
-        """
-        Get the global options.
+        """Get the global options.
 
         Returns:
             The global aria2c options.
@@ -645,10 +593,9 @@ class API:
         return Options(self, self.client.get_global_option())
 
     def set_options(self, options: OptionsType, downloads: list[Download]) -> list[bool]:
-        """
-        Set options for specific downloads.
+        """Set options for specific downloads.
 
-        Arguments:
+        Parameters:
             options: An instance of the [`Options`][aria2p.options.Options] class or a dictionary
                 containing aria2c options to create the download with.
             downloads: The list of downloads to set the options for.
@@ -656,10 +603,7 @@ class API:
         Returns:
             Success or failure of the operation for changing options for each given download.
         """
-        if isinstance(options, Options):
-            client_options = options.get_struct()
-        else:
-            client_options = options
+        client_options = options.get_struct() if isinstance(options, Options) else options
 
         # Note: batch/multicall candidate
         results = []
@@ -668,41 +612,35 @@ class API:
         return results
 
     def set_global_options(self, options: OptionsType) -> bool:
-        """
-        Set global options.
+        """Set global options.
 
-        Arguments:
+        Parameters:
             options: An instance of the [`Options`][aria2p.options.Options] class or a dictionary
                 containing aria2c options to create the download with.
 
         Returns:
             Success or failure of the operation for changing global options.
         """
-        if isinstance(options, Options):
-            client_options = options.get_struct()
-        else:
-            client_options = options
+        client_options = options.get_struct() if isinstance(options, Options) else options
 
         return self.client.change_global_option(client_options) == "OK"
 
     def get_stats(self) -> Stats:
-        """
-        Get the stats of the remote aria2c process.
+        """Get the stats of the remote aria2c process.
 
         Returns:
             The global stats returned by the remote process.
         """
         return Stats(self.client.get_global_stat())
 
-    @staticmethod  # noqa: WPS231,WPS602 (complex method, static method)
-    def remove_files(  # noqa: WPS231,WPS602
+    @staticmethod
+    def remove_files(
         downloads: list[Download],
-        force: bool = False,
+        force: bool = False,  # noqa: FBT001,FBT002
     ) -> list[bool]:
-        """
-        Remove downloaded files.
+        """Remove downloaded files.
 
-        Arguments:
+        Parameters:
             downloads:  the list of downloads for which to remove files.
             force: Whether to remove files even if download is not complete.
 
@@ -725,7 +663,7 @@ class API:
                     else:
                         try:
                             path.unlink()
-                        except FileNotFoundError as error:  # noqa: WPS440
+                        except FileNotFoundError as error:
                             logger.warning(f"File '{path}' did not exist when trying to delete it")
                             logger.opt(exception=True).trace(error)
                         results.append(True)
@@ -733,16 +671,15 @@ class API:
                 results.append(False)
         return results
 
-    @staticmethod  # noqa: WPS602 (static method)
-    def move_files(  # noqa: WPS602
+    @staticmethod
+    def move_files(
         downloads: list[Download],
         to_directory: PathOrStr,
-        force: bool = False,
+        force: bool = False,  # noqa: FBT001,FBT002
     ) -> list[bool]:
-        """
-        Move downloaded files to another directory.
+        """Move downloaded files to another directory.
 
-        Arguments:
+        Parameters:
             downloads:  the list of downloads for which to move files.
             to_directory: The target directory to move files to.
             force: Whether to move files even if download is not complete.
@@ -766,16 +703,15 @@ class API:
                 results.append(False)
         return results
 
-    @staticmethod  # noqa: WPS231,WPS602 (not that complex, static method)
-    def copy_files(  # noqa: WPS231,WPS602 (not that complex)
+    @staticmethod
+    def copy_files(
         downloads: list[Download],
         to_directory: PathOrStr,
-        force: bool = False,
+        force: bool = False,  # noqa: FBT001,FBT002
     ) -> list[bool]:
-        """
-        Copy downloaded files to another directory.
+        """Copy downloaded files to another directory.
 
-        Arguments:
+        Parameters:
             downloads:  the list of downloads for which to move files.
             to_directory: The target directory to copy files into.
             force: Whether to move files even if download is not complete.
@@ -805,7 +741,7 @@ class API:
 
     def listen_to_notifications(
         self,
-        threaded: bool = False,
+        threaded: bool = False,  # noqa: FBT001,FBT002
         on_download_start: Callable | None = None,
         on_download_pause: Callable | None = None,
         on_download_stop: Callable | None = None,
@@ -813,10 +749,9 @@ class API:
         on_download_error: Callable | None = None,
         on_bt_download_complete: Callable | None = None,
         timeout: int = 5,
-        handle_signals: bool = True,
+        handle_signals: bool = True,  # noqa: FBT001,FBT002
     ) -> None:
-        """
-        Start listening to aria2 notifications via WebSocket.
+        """Start listening to aria2 notifications via WebSocket.
 
         This method differs from [`Client.listen_to_notifications`][aria2p.client.Client.listen_to_notifications]
         in that it expects callbacks accepting two arguments, `api` and `gid`, instead of only `gid`.
@@ -824,7 +759,7 @@ class API:
 
         Stop listening to notifications with the [`API.stop_listening`][aria2p.api.API.stop_listening] method.
 
-        Arguments:
+        Parameters:
             threaded: Whether to start the listening loop in a thread or not (non-blocking or blocking).
             on_download_start: Callback for the `onDownloadStart` event.
             on_download_pause: Callback for the `onDownloadPause` event.
@@ -837,7 +772,7 @@ class API:
             handle_signals: Whether to add signal handlers to gracefully stop the loop on SIGTERM and SIGINT.
         """
 
-        def closure(callback):  # noqa: WPS430 (nested function)
+        def closure(callback: Callable) -> Callable:
             return functools.partial(callback, self) if callable(callback) else None
 
         kwargs = {
@@ -859,8 +794,7 @@ class API:
             self.client.listen_to_notifications(**kwargs)
 
     def stop_listening(self) -> None:
-        """
-        Stop listening to notifications.
+        """Stop listening to notifications.
 
         If the listening loop was threaded, this method will wait for the thread to finish.
         The time it takes for the thread to finish will depend on the timeout given while calling
@@ -871,11 +805,10 @@ class API:
             self.listener.join()
         self.listener = None
 
-    def split_input_file(self, lines: list[str] | TextIO) -> Iterator[list[str]]:  # noqa: WPS231 (not complex)
-        """
-        Helper to split downloads in an input file.
+    def split_input_file(self, lines: list[str] | TextIO) -> Iterator[list[str]]:
+        """Helper to split downloads in an input file.
 
-        Arguments:
+        Parameters:
              lines: The lines of the input file.
 
         Yields:
@@ -887,19 +820,17 @@ class API:
                 continue
             if not line.strip():  # Ignore empty line
                 continue
-            if not line.startswith(" "):  # URIs line
-                if block:
-                    yield block
-                    block = []
+            if not line.startswith(" ") and block:  # URIs line
+                yield block
+                block = []
             block.append(line.rstrip("\n"))
         if block:
             yield block
 
     def parse_input_file(self, input_file: PathOrStr) -> InputFileContentsType:
-        """
-        Parse a file with URIs or an aria2c input file.
+        """Parse a file with URIs or an aria2c input file.
 
-        Arguments:
+        Parameters:
             input_file: Path to file with URIs or aria2c input file.
 
         Returns:
@@ -910,12 +841,12 @@ class API:
             for download_lines in self.split_input_file(fd):
                 uris = download_lines[0].split("\t")
                 options = {}
-                try:  # noqa: WPS229
+                try:
                     for option_line in download_lines[1:]:
                         option_name, option_value = option_line.split("=", 1)
                         options[option_name.lstrip()] = option_value
                     downloads.append((uris, options))
                 except ValueError as error:
-                    logger.error(f"Skipping download because of invalid option line '{option_line}'")  # noqa: WPS441
+                    logger.error(f"Skipping download because of invalid option line '{option_line}'")
                     logger.opt(exception=True).trace(error)
         return downloads
