@@ -3,9 +3,12 @@
 from __future__ import annotations
 
 import sys
+from typing import TYPE_CHECKING
 
-from aria2p.api import API
 from aria2p.utils import read_lines
+
+if TYPE_CHECKING:
+    from aria2p.api import API
 
 
 def add_magnets(
@@ -15,10 +18,9 @@ def add_magnets(
     options: dict | None = None,
     position: int | None = None,
 ) -> int:
-    """
-    Add magnet subcommand.
+    """Add magnet subcommand.
 
-    Arguments:
+    Parameters:
         api: The API instance to use.
         uris: The URIs of the magnets.
         from_file: Path to the file to read uris from.
