@@ -2,14 +2,16 @@
 
 from __future__ import annotations
 
-from aria2p.api import API
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from aria2p.api import API
 
 
 def purge(api: API) -> int:
-    """
-    Purge subcommand.
+    """Purge subcommand.
 
-    Arguments:
+    Parameters:
         api: The API instance to use.
 
     Returns:
