@@ -3,16 +3,23 @@
 from __future__ import annotations
 
 import sys
+from typing import TYPE_CHECKING
 
-from aria2p.api import API
 from aria2p.client import ClientException
 
+if TYPE_CHECKING:
+    from aria2p.api import API
 
-def remove(api: API, gids: list[str] | None = None, do_all: bool = False, force: bool = False) -> int:
-    """
-    Remove subcommand.
 
-    Arguments:
+def remove(
+    api: API,
+    gids: list[str] | None = None,
+    do_all: bool = False,  # noqa: FBT001,FBT002
+    force: bool = False,  # noqa: FBT001,FBT002
+) -> int:
+    """Remove subcommand.
+
+    Parameters:
         api: The API instance to use.
         gids: The GIDs of the downloads to remove.
         do_all: Pause all downloads if True.
