@@ -1,8 +1,14 @@
+"""Fake HTTP server."""
+
+from __future__ import annotations
+
+from typing import AsyncIterator
+
 from fastapi import FastAPI
 from fastapi.responses import StreamingResponse
 
 
-def translate_size(size):
+def translate_size(size: str) -> int:
     try:
         return int(size)
     except ValueError:
@@ -19,7 +25,7 @@ def translate_size(size):
     return int(size.rstrip("kmg")) * multiplier
 
 
-async def virtual_file(size, chunks=4096):
+async def virtual_file(size: int, chunks: int = 4096) -> AsyncIterator[bytes]:
     while size > 0:
         yield b"1" * min(size, chunks)
         size -= chunks
@@ -29,7 +35,7 @@ app = FastAPI()
 
 
 @app.get("/{size}")
-async def get(size):
+async def get(size: str) -> StreamingResponse:
     return StreamingResponse(
         virtual_file(translate_size(size)),
         media_type="application/octet-stream",
