@@ -1,38 +1,43 @@
 """Tests for the `api` module."""
 
+from __future__ import annotations
+
 import threading
 import time
+from typing import TYPE_CHECKING
 
 import pytest
 
 from aria2p import API, Client, ClientException, Download
+from tests import BUNSENLABS_MAGNET, BUNSENLABS_TORRENT, CONFIGS_DIR, DEBIAN_METALINK, INPUT_FILES, XUBUNTU_MIRRORS
+from tests.conftest import Aria2Server
 
-from . import BUNSENLABS_MAGNET, BUNSENLABS_TORRENT, CONFIGS_DIR, DEBIAN_METALINK, INPUT_FILES, XUBUNTU_MIRRORS
-from .conftest import Aria2Server
+if TYPE_CHECKING:
+    from pathlib import Path
 
 
-def test_add_magnet_method(server):
+def test_add_magnet_method(server: Aria2Server) -> None:
     assert server.api.add_magnet(BUNSENLABS_MAGNET)
 
 
-def test_add_metalink_method(server):
+def test_add_metalink_method(server: Aria2Server) -> None:
     assert server.api.add_metalink(DEBIAN_METALINK)
 
 
-def test_add_torrent_method(server):
+def test_add_torrent_method(server: Aria2Server) -> None:
     assert server.api.add_torrent(BUNSENLABS_TORRENT)
 
 
-def test_add_uris_method(server):
+def test_add_uris_method(server: Aria2Server) -> None:
     assert server.api.add_uris(XUBUNTU_MIRRORS)
 
 
-def test_get_download_method(tmp_path, port):
+def test_get_download_method(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, session="1-dl-paused.txt") as server:
         assert server.api.get_download("0000000000000001")  # == server.api.get_downloads()[0].gid
 
 
-def test_get_downloads_method(tmp_path, port):
+def test_get_downloads_method(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, session="2-dls.txt") as server:
         downloads = server.api.get_downloads()
         assert len(downloads) == 2
@@ -40,25 +45,25 @@ def test_get_downloads_method(tmp_path, port):
         assert downloads[0].gid == "0000000000000001"
 
 
-def test_get_global_options_method(tmp_path, port):
+def test_get_global_options_method(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, config=CONFIGS_DIR / "max-5-dls.conf") as server:
         options = server.api.get_global_options()
         assert options.download is None
         assert options.max_concurrent_downloads == 5
 
 
-def test_get_options_method(tmp_path, port):
+def test_get_options_method(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, session="max-dl-limit-10000.txt") as server:
         downloads = server.api.get_downloads()
         options = server.api.get_options(downloads)[0]
         assert options.max_download_limit == 10000
 
 
-def test_get_stats_method(server):
+def test_get_stats_method(server: Aria2Server) -> None:
     assert server.api.get_stats()
 
 
-def test_move_method(tmp_path, port):
+def test_move_method(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, session="2-dls-paused.txt") as server:
         downloads = server.api.get_downloads()
         assert server.api.move(downloads[0], 1) == 1
@@ -66,7 +71,7 @@ def test_move_method(tmp_path, port):
         assert downloads == list(reversed(new_pos_downloads))
 
 
-def test_move_down_method(tmp_path, port):
+def test_move_down_method(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, session="2-dls-paused.txt") as server:
         downloads = server.api.get_downloads()
         assert server.api.move_down(downloads[0]) == 1
@@ -74,7 +79,7 @@ def test_move_down_method(tmp_path, port):
         assert downloads == list(reversed(new_pos_downloads))
 
 
-def test_move_to_method(tmp_path, port):
+def test_move_to_method(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, session="2-dls-paused.txt") as server:
         downloads = server.api.get_downloads()
         assert server.api.move_to(downloads[0], 1) == 1
@@ -86,7 +91,7 @@ def test_move_to_method(tmp_path, port):
         assert new_pos_downloads == downloads
 
 
-def test_move_to_bottom_method(tmp_path, port):
+def test_move_to_bottom_method(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, session="2-dls-paused.txt") as server:
         downloads = server.api.get_downloads()
         assert server.api.move_to_bottom(downloads[0]) == 1
@@ -94,7 +99,7 @@ def test_move_to_bottom_method(tmp_path, port):
         assert new_pos_downloads == [downloads[1], downloads[0]]
 
 
-def test_move_to_top_method(tmp_path, port):
+def test_move_to_top_method(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, session="2-dls-paused.txt") as server:
         downloads = server.api.get_downloads()
         assert server.api.move_to_top(downloads[1]) == 0
@@ -102,7 +107,7 @@ def test_move_to_top_method(tmp_path, port):
         assert new_pos_downloads == [downloads[1], downloads[0]]
 
 
-def test_move_up_method(tmp_path, port):
+def test_move_up_method(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, session="2-dls-paused.txt") as server:
         downloads = server.api.get_downloads()
         assert server.api.move_up(downloads[0]) == 0
@@ -110,7 +115,7 @@ def test_move_up_method(tmp_path, port):
         assert downloads == new_pos_downloads
 
 
-def test_pause_method(tmp_path, port):
+def test_pause_method(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, session="big-download.txt") as server:
         time.sleep(0.1)
         downloads = server.api.get_downloads()
@@ -120,7 +125,7 @@ def test_pause_method(tmp_path, port):
         assert downloads[0].live.status == "paused"
 
 
-def test_pause_all_method(tmp_path, port):
+def test_pause_all_method(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, session="3-dls.txt") as server:
         assert server.api.pause_all()
         # The following code block is commented out because we cannot ensure
@@ -142,12 +147,12 @@ def test_pause_all_method(tmp_path, port):
         #     raise AssertionError
 
 
-def test_autopurge_method(tmp_path, port):
+def test_autopurge_method(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, session="3-dls.txt") as server:
         assert server.api.autopurge()
 
 
-def test_remove_method(tmp_path, port):
+def test_remove_method(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, session="3-dls.txt") as server:
         downloads = server.api.get_downloads()
         if not all(server.api.remove(downloads)):
@@ -156,7 +161,7 @@ def test_remove_method(tmp_path, port):
         assert not downloads
 
 
-def test_remove_files_method(tmp_path, port):
+def test_remove_files_method(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, session="very-small-download.txt") as server:
         time.sleep(1)
         download = server.api.get_downloads()[0]
@@ -169,7 +174,7 @@ def test_remove_files_method(tmp_path, port):
             assert not file.exists()
 
 
-def test_remove_files_not_complete(tmp_path, port):
+def test_remove_files_not_complete(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, session="2-dls-paused.txt") as server:
         downloads = server.api.get_downloads()
         assert server.api.remove(downloads, files=True)
@@ -178,19 +183,19 @@ def test_remove_files_not_complete(tmp_path, port):
                 assert file.exists()
 
 
-def test_remove_files_tree(server):
+def test_remove_files_tree(server: Aria2Server) -> None:
     directory = server.tmp_dir / "some-directory"
     directory.mkdir()
 
     class _Download:
         is_complete = True
-        root_files_paths = [directory]
+        root_files_paths = [directory]  # noqa: RUF012
 
     assert server.api.remove_files([_Download])
     assert not directory.exists()
 
 
-def test_remove_all_method(tmp_path, port):
+def test_remove_all_method(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, session="3-dls.txt") as server:
         if not server.api.remove_all():
             pytest.xfail("Sporadic failures")
@@ -198,15 +203,15 @@ def test_remove_all_method(tmp_path, port):
         assert not downloads
 
 
-def test_resume_method(tmp_path, port):
+def test_resume_method(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, session="2-dls-paused.txt") as server:
         downloads = server.api.get_downloads()
         assert server.api.resume(downloads)
         downloads = server.api.get_downloads()
-        assert all([d.is_active for d in downloads])
+        assert all(d.is_active for d in downloads)
 
 
-def test_resume_all_method(tmp_path, port):
+def test_resume_all_method(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, session="2-dls-paused.txt") as server:
         time.sleep(0.1)
         assert server.api.resume_all()
@@ -215,17 +220,17 @@ def test_resume_all_method(tmp_path, port):
         for download in downloads:
             if download.has_failed:
                 pytest.xfail("Failed to establish connection (sporadic error)")
-        assert all([d.is_active for d in downloads])
+        assert all(d.is_active for d in downloads)
 
 
-def test_set_global_options_method(tmp_path, port):
+def test_set_global_options_method(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, config=CONFIGS_DIR / "max-5-dls.conf") as server:
         assert server.api.set_global_options({"max-concurrent-downloads": "10"})
         options = server.api.get_global_options()
         assert options.max_concurrent_downloads == 10
 
 
-def test_set_options_method(tmp_path, port):
+def test_set_options_method(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, session="max-dl-limit-10000.txt") as server:
         downloads = server.api.get_downloads()
         try:
@@ -240,7 +245,7 @@ def test_set_options_method(tmp_path, port):
 
 
 # @pytest.mark.flaky(reruns=5)
-def test_copy_files_method(tmp_path_factory, port):
+def test_copy_files_method(tmp_path_factory: pytest.TmpPathFactory, port: int) -> None:
     with Aria2Server(tmp_path_factory.mktemp("copy_files"), port, session="very-small-download.txt") as server:
         # initialize temp dir to copy to
         tmp_dir = tmp_path_factory.mktemp("copy_files")
@@ -273,7 +278,7 @@ def test_copy_files_method(tmp_path_factory, port):
         tmp_dir.rmdir()
 
 
-def test_move_files_method(tmp_path_factory, port):
+def test_move_files_method(tmp_path_factory: pytest.TmpPathFactory, port: int) -> None:
     with Aria2Server(tmp_path_factory.mktemp("move_files"), port, session="very-small-download.txt") as server:
         # initialize temp dir to copy to
         tmp_dir = tmp_path_factory.mktemp("move_files")
@@ -308,24 +313,26 @@ def test_move_files_method(tmp_path_factory, port):
         tmp_dir.rmdir()
 
 
-def test_listen_to_notifications(tmp_path, port):
+def test_listen_to_notifications(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, session="2-dls-paused.txt") as server:
         server.api.listen_to_notifications(threaded=True, timeout=1)
     time.sleep(3)
     assert not server.api.listener.is_alive()
 
 
-def test_listen_to_notifications_then_stop(port):
+def test_listen_to_notifications_then_stop(port: int) -> None:
     api = API(Client(port=port))
     api.listen_to_notifications(threaded=True, timeout=1)
     api.stop_listening()
     assert api.listener is None
 
 
-def test_listen_to_notifications_callbacks(tmp_path, port, capsys):
+def test_listen_to_notifications_callbacks(tmp_path: Path, port: int, capsys: pytest.CaptureFixture) -> None:
     with Aria2Server(tmp_path, port, session="2-dls-paused.txt") as server:
         server.api.listen_to_notifications(
-            on_download_start=lambda api, gid: print("started " + gid), threaded=True, timeout=1
+            on_download_start=lambda api, gid: print("started " + gid),  # noqa: T201
+            threaded=True,
+            timeout=1,
         )
         time.sleep(1)
         server.api.resume_all()
@@ -334,10 +341,10 @@ def test_listen_to_notifications_callbacks(tmp_path, port, capsys):
     assert capsys.readouterr().out == "started 0000000000000001\nstarted 0000000000000002\n"
 
 
-def test_listen_to_notifications_no_thread(tmp_path, port):
+def test_listen_to_notifications_no_thread(tmp_path: Path, port: int) -> None:
     with Aria2Server(tmp_path, port, session="2-dls-paused.txt") as server:
 
-        def thread_target():
+        def thread_target() -> None:
             server.api.listen_to_notifications(threaded=False, timeout=1)
 
         thread = threading.Thread(target=thread_target)
@@ -348,7 +355,7 @@ def test_listen_to_notifications_no_thread(tmp_path, port):
         server.api.stop_listening()
 
 
-def test_parse_input_file():
+def test_parse_input_file() -> None:
     api = API()
 
     downloads = api.parse_input_file(INPUT_FILES[0])
