@@ -1,9 +1,12 @@
 """Tests for the `interface` module."""
 
+from __future__ import annotations
+
 import os
 import sys
 import time
 from pathlib import Path
+from typing import TYPE_CHECKING, Any, Callable
 
 import pyperclip
 import pytest
@@ -11,20 +14,30 @@ from asciimatics.event import KeyboardEvent, MouseEvent
 from asciimatics.screen import Screen
 
 from aria2p import interface as tui
+from tests import TESTS_DATA_DIR
+from tests.conftest import Aria2Server
 
-from . import TESTS_DATA_DIR
-from .conftest import Aria2Server
+if TYPE_CHECKING:
+    from aria2p.api import API
 
 tui.Interface.frames = 20  # reduce tests time
 
 
 class SpecialEvent:
+    """Special events for special actions."""
+
     RESIZE = 1
     PASS_N_FRAMES = 2
     PASS_N_TICKS = 3
     RAISE = 4
 
-    def __init__(self, event_type, value=None):
+    def __init__(self, event_type: int, value: Any = None) -> None:
+        """Initialize the special event.
+
+        Parameters:
+            event_type: The event type (enumeration value).
+            value: The event value (could be anything).
+        """
         self.type = event_type
         self.value = value
 
@@ -57,23 +70,29 @@ class Event:
     space = KeyboardEvent(ord(" "))
 
     @staticmethod
-    def hit(letter):
+    def hit(letter: str) -> KeyboardEvent:
         return KeyboardEvent(ord(letter))
 
     @staticmethod
-    def exc(value):
+    def exc(value: type[Exception]) -> SpecialEvent:
         return SpecialEvent(SpecialEvent.RAISE, value)
 
     @staticmethod
-    def pass_frames(value):
+    def pass_frames(value: int) -> SpecialEvent:
         return SpecialEvent(SpecialEvent.PASS_N_FRAMES, value)
 
     @staticmethod
-    def pass_ticks(value):
+    def pass_ticks(value: int) -> SpecialEvent:
         return SpecialEvent(SpecialEvent.PASS_N_TICKS, value)
 
 
-def get_interface(patcher, api=None, events=None, append_q=True):
+def get_interface(
+    patcher: Callable,
+    api: API | None = None,
+    events: list[Event] | None = None,
+    *,
+    append_q: bool = True,
+) -> tui.Interface:
     if not events:
         events = []
 
@@ -84,14 +103,21 @@ def get_interface(patcher, api=None, events=None, append_q=True):
         def __enter__(self):
             return MockedScreen(events)
 
-        def __exit__(self, exc_type, exc_val, exc_tb):
+        def __exit__(self, exc_type, exc_val, exc_tb):  # noqa: ANN001
             pass
 
     patcher.setattr(tui, "ManagedScreen", MockedManagedScreen)
     return tui.Interface(api=api)
 
 
-def run_interface(patcher, api=None, events=None, append_q=True, **kwargs):
+def run_interface(
+    patcher: Callable,
+    api: API | None = None,
+    events: list[Event] | None = None,
+    *,
+    append_q: bool = True,
+    **kwargs: Any,
+) -> tui.Interface:
     interface = get_interface(patcher, api, events, append_q)
     for key, value in kwargs.items():
         setattr(interface, key, value)
@@ -100,7 +126,12 @@ def run_interface(patcher, api=None, events=None, append_q=True, **kwargs):
 
 
 class MockedScreen:
-    def __init__(self, events):
+    def __init__(self, events: list[Event]) -> None:
+        """Initialize the mocked screen.
+
+        Parameters:
+            events: Pre-recorded events.
+        """
         self.events = events
         self._has_resized = False
         self._pass_n_frames = 0
@@ -109,26 +140,26 @@ class MockedScreen:
         self.n_refresh = 0
 
     @property
-    def dimensions(self):
+    def dimensions(self) -> tuple[int, int]:
         return 30, 80
 
-    def has_resized(self):
+    def has_resized(self) -> bool:
         return self._has_resized
 
-    def open(self):
+    def open(self) -> None:  # noqa: A003
         pass
 
-    def close(self):
+    def close(self) -> None:
         pass
 
-    def get_event(self):
+    def get_event(self) -> KeyboardEvent | MouseEvent | None:
         if self._pass_n_frames > 0:
             self._pass_n_frames -= 1
             return None
         event = self.events.pop(0)
         if isinstance(event, (KeyboardEvent, MouseEvent)):
             return event
-        elif isinstance(event, SpecialEvent):
+        if isinstance(event, SpecialEvent):
             if event.type == SpecialEvent.RESIZE:
                 self._has_resized = True
             elif event.type == SpecialEvent.PASS_N_FRAMES:
@@ -141,38 +172,38 @@ class MockedScreen:
                 raise event.value
         return None
 
-    def print_at(self, *args, **kwargs):
+    def print_at(self, *args: Any, **kwargs: Any) -> None:
         if args[0].strip():
             self.print_at_calls.append({"args": args, "kwargs": kwargs})
 
-    def paint(self, *args, **kwargs):
+    def paint(self, *args: Any, **kwargs: Any) -> None:
         if args[0].strip():
             self.paint_calls.append({"args": args, "kwargs": kwargs})
 
-    def refresh(self):
+    def refresh(self) -> None:
         self.n_refresh += 1
 
 
-def test_run(monkeypatch):
+def test_run(monkeypatch: pytest.MonkeyPatch) -> None:
     interface = run_interface(monkeypatch)
     assert interface.screen
     assert interface.height == 30
     assert interface.width == 80
 
 
-def test_resize(server, monkeypatch):
+def test_resize(server: Aria2Server, monkeypatch: pytest.MonkeyPatch) -> None:
     interface = run_interface(monkeypatch, server.api, events=[Event.resize])
     # assert screen was renewed
     assert not interface.screen.has_resized()
 
 
-def test_frames_plus_n(server, monkeypatch):
+def test_frames_plus_n(server: Aria2Server, monkeypatch: pytest.MonkeyPatch) -> None:
     n = 10
     interface = run_interface(monkeypatch, server.api, events=[Event.pass_frames(tui.Interface.frames + n)])
     assert interface.frame == n
 
 
-def test_change_sort(server, monkeypatch):
+def test_change_sort(server: Aria2Server, monkeypatch: pytest.MonkeyPatch) -> None:
     interface = run_interface(
         monkeypatch,
         server.api,
@@ -188,7 +219,7 @@ def test_change_sort(server, monkeypatch):
     assert interface.sort == tui.Interface.sort - 1
 
 
-def test_move_focus(tmp_path, port, monkeypatch):
+def test_move_focus(tmp_path: Path, port: int, monkeypatch: pytest.MonkeyPatch) -> None:
     with Aria2Server(tmp_path, port, session="2-dls-paused.txt") as server:
         interface = run_interface(
             monkeypatch,
@@ -198,12 +229,12 @@ def test_move_focus(tmp_path, port, monkeypatch):
     assert interface.focused == 0
 
 
-def test_show_help(server, monkeypatch):
+def test_show_help(server: Aria2Server, monkeypatch: pytest.MonkeyPatch) -> None:
     interface = run_interface(monkeypatch, server.api, events=[Event.f1, Event.pass_tick, Event.enter])
     assert interface.screen.print_at_calls[-1]["args"][0].startswith("Press any key to return.")
 
 
-def test_horizontal_scrolling(tmp_path, port, monkeypatch):
+def test_horizontal_scrolling(tmp_path: Path, port: int, monkeypatch: pytest.MonkeyPatch) -> None:
     with Aria2Server(tmp_path, port, session="3-magnets.txt") as server:
         run_interface(
             monkeypatch,
@@ -218,7 +249,7 @@ def test_horizontal_scrolling(tmp_path, port, monkeypatch):
         )
 
 
-def test_log_exception(tmp_path, port, monkeypatch):
+def test_log_exception(tmp_path: Path, port: int, monkeypatch: pytest.MonkeyPatch) -> None:
     with Aria2Server(tmp_path, port, session="2-dls-paused.txt") as server:
         interface = get_interface(monkeypatch, server.api, events=[Event.exc(LookupError("some message"))])
         assert not interface.run()
@@ -234,7 +265,7 @@ def test_log_exception(tmp_path, port, monkeypatch):
     assert "LookupError" in last_line
 
 
-def test_select_sort(tmp_path, port, monkeypatch):
+def test_select_sort(tmp_path: Path, port: int, monkeypatch: pytest.MonkeyPatch) -> None:
     with Aria2Server(tmp_path, port, session="3-magnets.txt") as server:
         run_interface(
             monkeypatch,
@@ -252,7 +283,7 @@ def test_select_sort(tmp_path, port, monkeypatch):
         )
 
 
-def test_mouse_event(tmp_path, port, monkeypatch):
+def test_mouse_event(tmp_path: Path, port: int, monkeypatch: pytest.MonkeyPatch) -> None:
     reverse = tui.Interface.reverse
     with Aria2Server(tmp_path, port, session="3-magnets.txt") as server:
         interface = run_interface(
@@ -264,12 +295,12 @@ def test_mouse_event(tmp_path, port, monkeypatch):
     assert interface.reverse is not reverse
 
 
-def test_vertical_scrolling(tmp_path, port, monkeypatch):
+def test_vertical_scrolling(tmp_path: Path, port: int, monkeypatch: pytest.MonkeyPatch) -> None:
     with Aria2Server(tmp_path, port, session="50-dls.txt") as server:
         run_interface(monkeypatch, server.api, events=[Event.pass_frame] + [Event.down] * 40 + [Event.up] * 30)
 
 
-def test_follow_focused(tmp_path, port, monkeypatch):
+def test_follow_focused(tmp_path: Path, port: int, monkeypatch: pytest.MonkeyPatch) -> None:
     with Aria2Server(tmp_path, port, session="3-magnets.txt") as server:
         interface = run_interface(
             monkeypatch,
@@ -280,7 +311,7 @@ def test_follow_focused(tmp_path, port, monkeypatch):
     assert interface.focused == 2
 
 
-def test_remove_ask(tmp_path, port, monkeypatch):
+def test_remove_ask(tmp_path: Path, port: int, monkeypatch: pytest.MonkeyPatch) -> None:
     with Aria2Server(tmp_path, port, session="very-small-download.txt") as server:
         download = server.api.get_downloads()[0]
         while not download.live.is_complete:
@@ -297,12 +328,12 @@ def test_remove_ask(tmp_path, port, monkeypatch):
         assert interface.last_remove_choice == 1
 
 
-def test_setup(tmp_path, port, monkeypatch):
+def test_setup(tmp_path: Path, port: int, monkeypatch: pytest.MonkeyPatch) -> None:
     with Aria2Server(tmp_path, port, session="2-dls-paused.txt") as server:
         run_interface(monkeypatch, server.api, events=[Event.pass_frame, Event.f2, Event.pass_tick])
 
 
-def test_toggle_resume_pause(tmp_path, port, monkeypatch):
+def test_toggle_resume_pause(tmp_path: Path, port: int, monkeypatch: pytest.MonkeyPatch) -> None:
     with Aria2Server(tmp_path, port, session="big-download.txt") as server:
         interface = run_interface(
             monkeypatch,
@@ -326,7 +357,7 @@ def test_toggle_resume_pause(tmp_path, port, monkeypatch):
     assert interface.data[0].is_paused
 
 
-def test_priority(tmp_path, port, monkeypatch):
+def test_priority(tmp_path: Path, port: int, monkeypatch: pytest.MonkeyPatch) -> None:
     with Aria2Server(tmp_path, port, session="2-dls-paused.txt") as server:
         run_interface(
             monkeypatch,
@@ -344,13 +375,13 @@ def test_priority(tmp_path, port, monkeypatch):
         )
 
 
-def test_sort_edges(tmp_path, port, monkeypatch):
+def test_sort_edges(tmp_path: Path, port: int, monkeypatch: pytest.MonkeyPatch) -> None:
     with Aria2Server(tmp_path, port, session="2-dls-paused.txt") as server:
         run_interface(monkeypatch, server.api, events=[Event.hit("<")], sort=0)
         run_interface(monkeypatch, server.api, events=[Event.hit(">")], sort=7)
 
 
-def test_remember_last_remove(tmp_path, port, monkeypatch):
+def test_remember_last_remove(tmp_path: Path, port: int, monkeypatch: pytest.MonkeyPatch) -> None:
     with Aria2Server(tmp_path, port, session="2-dls-paused.txt") as server:
         interface = run_interface(
             monkeypatch,
@@ -361,7 +392,7 @@ def test_remember_last_remove(tmp_path, port, monkeypatch):
     assert interface.side_focused == 1
 
 
-def test_autoclear(tmp_path, port, monkeypatch):
+def test_autoclear(tmp_path: Path, port: int, monkeypatch: pytest.MonkeyPatch) -> None:
     with Aria2Server(tmp_path, port, session="very-small-download.txt") as server:
         download = server.api.get_downloads()[0]
         while not download.live.is_complete:
@@ -372,7 +403,7 @@ def test_autoclear(tmp_path, port, monkeypatch):
     assert not interface.data
 
 
-def test_side_column_edges(tmp_path, port, monkeypatch):
+def test_side_column_edges(tmp_path: Path, port: int, monkeypatch: pytest.MonkeyPatch) -> None:
     with Aria2Server(tmp_path, port, session="2-dls-paused.txt") as server:
         run_interface(
             monkeypatch,
@@ -400,17 +431,21 @@ def test_side_column_edges(tmp_path, port, monkeypatch):
         )
 
 
-def test_click_row(tmp_path, port, monkeypatch):
+def test_click_row(tmp_path: Path, port: int, monkeypatch: pytest.MonkeyPatch) -> None:
     with Aria2Server(tmp_path, port, session="2-dls-paused.txt") as server:
         interface = run_interface(
-            monkeypatch, server.api, events=[Event.pass_frame, MouseEvent(x=10, y=2, buttons=MouseEvent.LEFT_CLICK)]
+            monkeypatch,
+            server.api,
+            events=[Event.pass_frame, MouseEvent(x=10, y=2, buttons=MouseEvent.LEFT_CLICK)],
         )
     assert interface.focused == 1
 
 
-def test_click_out_bounds(server, monkeypatch):
+def test_click_out_bounds(server: Aria2Server, monkeypatch: pytest.MonkeyPatch) -> None:
     run_interface(
-        monkeypatch, server.api, events=[Event.pass_frame, MouseEvent(x=1000, y=0, buttons=MouseEvent.LEFT_CLICK)]
+        monkeypatch,
+        server.api,
+        events=[Event.pass_frame, MouseEvent(x=1000, y=0, buttons=MouseEvent.LEFT_CLICK)],
     )
     with open(Path("tests") / "logs" / "test_interface" / "test_click_out_bounds.log") as log_file:
         lines = log_file.readlines()
@@ -427,7 +462,7 @@ def test_click_out_bounds(server, monkeypatch):
     sys.platform != "win32" and "CI" in os.environ,
     reason="Can't get pyperclip to work on Linux or MacOS in CI",
 )
-def test_add_downloads_uris(server, monkeypatch):
+def test_add_downloads_uris(server: Aria2Server, monkeypatch: pytest.MonkeyPatch) -> None:
     clipboard_selection_downloads = ""
     primary_selection_downloads = ""
 
@@ -470,7 +505,7 @@ def test_add_downloads_uris(server, monkeypatch):
     sys.platform != "win32" and "CI" in os.environ,
     reason="Can't get pyperclip to work on Linux or MacOS in CI",
 )
-def test_add_downloads_torrents_and_metalinks(server, monkeypatch):
+def test_add_downloads_torrents_and_metalinks(server: Aria2Server, monkeypatch: pytest.MonkeyPatch) -> None:
     torrent_file = TESTS_DATA_DIR / "bunsenlabs-helium-4.iso.torrent"
     metalink_file = TESTS_DATA_DIR / "debian.metalink"
 
